% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTriplet.R
\name{createTriplet}
\alias{createTriplet}
\title{This function creates the triplet element for an item if the full triplet is
missing.}
\usage{
createTriplet(data, formula)
}
\arguments{
\item{data}{The data, must be denormalised}

\item{formula}{The formula table, see \code{getYieldFormula}.}
}
\value{
The original data if the triplet is present, otherwise a new
    data.table with the missing elements created.
}
\description{
The element triplet includes the input, productivity and output. However,
for certain commodities certain element was never collected and under the
sparse data representation the element is not availble.
}
\details{
This function creates the temporary element that is required for processing.
}
