% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandYear.R
\name{expandYear}
\alias{expandYear}
\title{This function expands the range of the year variable.}
\usage{
expandYear(
  data,
  areaVar = "geographicAreaM49",
  elementVar = "measuredElement",
  itemVar = "measuredItemCPC",
  yearVar = "timePointYears",
  valueVar = "Value",
  obsflagVar = "flagObservationStatus",
  methFlagVar = "flagMethod",
  newYears = NULL
)
}
\arguments{
\item{data}{The dataset}

\item{areaVar}{The column name corresponding to the geographic area.}

\item{elementVar}{The column name corresponds to the measured element.}

\item{itemVar}{The column name corresponding to the commodity item.}

\item{yearVar}{The column name corresponds to the time dimension.}

\item{valueVar}{The regular expression to capture the value column.}
}
\value{
A new dataset with the year variable expanded
}
\description{
This is required due the sparse representation of databases. When a new
round of processing commencing in 2015, since there are no data input nor
previous imputation, the record does not exist and thus imputation or any
processing has no record to replace. We need to expand the year variable in
order to perform imputation.
}
\details{
New records created are not filled with flags, both observation status and
method flag remains NA. The reason behind this is so that after the
processing, if the record can not be populated with imputation then it will
be dropped and not saved back to the database.
}
