% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDenormalisedElementCodes.R
\name{getDenormalisedElementCodes}
\alias{getDenormalisedElementCodes}
\title{Function to get all the elements of a denormalised dataset}
\usage{
getDenormalisedElementCodes(denormalisedData)
}
\arguments{
\item{denormalisedData}{A data.table with the elements denormalised.}
}
\value{
The element codes in the data
}
\description{
Function to get all the elements of a denormalised dataset
}
