% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{preProcessing}
\alias{preProcessing}
\title{This function performs manipulation of the data that are standard
after the data is retrieved from the data base.}
\usage{
preProcessing(data, normalised = TRUE, dropNonExistingRecord = TRUE,
  denormalisedKey = "measuredElement")
}
\arguments{
\item{data}{The data.table object}

\item{normalised}{logical, whether the data is normalised}

\item{dropNonExistingRecord}{logical, whether non existing records should be
droped. Non-existing records are those having flagObservationStatus and
flagMethod have the value NA. This can be generated from denormalising
the data or occassionaly returned by the GetData function.}

\item{denormalisedKey}{optional, only required if the input data is not
normalised.It is the name of the key that denormalises the data.}
}
\value{
A data.table with standard pre processing steps
    performed.
}
\description{
This function performs manipulation of the data that are standard
after the data is retrieved from the data base.
}

