% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCalculated.R
\name{removeCalculated}
\alias{removeCalculated}
\title{This function removes previous calculated values}
\usage{
removeCalculated(
  data,
  valueVar,
  observationFlagVar,
  methodFlagVar,
  missingObservationFlag = "M",
  missingMethodFlag = "u",
  calculatedMethodFlag = "i"
)
}
\arguments{
\item{data}{The data.table object containing the values.}

\item{valueVar}{The column name of data which contains the values for the
variable to be modified.}

\item{observationFlagVar}{The column name of the observation flag
corresponding to value.}

\item{methodFlagVar}{The column name of the method flag corresponding to
value.}

\item{missingObservationFlag}{The character value which is used to represent
missing values in the observation flag. This value will be put in any
column containing imputation data that is removed.}

\item{missingMethodFlag}{The character value which is used to represent
missing values in the method flag. This value will be put in any column
containing imputation data that is removed.}

\item{calculatedMethodFlag}{This character value specifies which method flag
correspond to previous calculated values in the data.}
}
\value{
No value is returned. However, the object "data" which was passed to
    this function is modified.
}
\description{
This function removes previous calculated values
}
\details{
The observation flag is not required, as previous calculated values
    can contain any observation flag due to flag aggregation.
}
