% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeInvalidFlag.R
\name{removeInvalidFlag}
\alias{removeInvalidFlag}
\title{This function has been built in order to transform all the values corresponding to ivalid flag combinations
into NA M u and activate the imputation process.}
\usage{
removeInvalidFlag(data, valueVar = "Value",
  observationFlagVar = "flagObservationStatus",
  methodFlagVar = "flagMethod", missingObservationFlag = "M",
  missingMethodFlag = "u", normalised = TRUE,
  denormalisedKey = "measuredElement", flagValidTable = NULL)
}
\arguments{
\item{data}{The data.table object to be manipulate}

\item{valueVar}{The column name of data which contains the values for the
variable to be modified.}

\item{observationFlagVar}{The column name of the observation flag
corresponding to value.}

\item{methodFlagVar}{The column name of the method flag corresponding to
value.}

\item{missingObservationFlag}{The character value which is used to represent
missing values in the observation flag. This value will be put in any
column containing imputation data that is removed.}

\item{missingMethodFlag}{The character value which is used to represent
missing values in the method flag. This value will be put in any column
containing imputation data that is removed.}

\item{normalised}{logical, whether the data is normalised}

\item{denormalisedKey}{optional, only required if the input data is not
normalised.It is the name of the key that denormalises the data.}

\item{flagValidTable}{Table containing the Valid/Invalid and Protected/Non
protected flag combinations}
}
\description{
This function has been built in order to transform all the values corresponding to ivalid flag combinations
into NA M u and activate the imputation process.
}
\details{
flagValidTable Table containing valid and protected flag combination.
 This table is generally recorded into the Flag package.
}

