% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faoswsProcessing-package.R
\docType{package}
\name{faoswsProcessing-package}
\alias{faoswsProcessing}
\alias{faoswsProcessing-package}
\title{Provide standardised processing procedures for data manipulation}
\description{
This package provides a family of functions to standardise data manipulation.
}
\author{
Michael C. J. Kao
}

